Extract most of page controllers to concern
