+++ app/controllers/concerns/pages.rb
@@ -1,0 +1,43 @@
+module Pages
+  extend ActiveSupport::Concern
+
+  included do
+    resource_actions :default
+
+    schema_type :strong_params
+
+    allowed_params :create, :url_key, :title, :content, :language
+    allowed_params :update, :url_key, :title, :content, :language
+
+    before_filter :set_language_from_header, only: [:index]
+
+    define_method(:link_header) do |resource|
+      resource = resource.first
+      send(:"api_#{ resource_name }_url", id: resource.id, "#{parent_resource}_id": resource_id)
+    end
+
+    define_method(:build_resource_for_create) do |create_params|
+      create_params[:links] ||= {}
+      create_params[:links][parent_resource] = resource_id
+      super create_params
+    end
+  end
+
+  def serializer
+    @serializer ||= "#{resource_name.camelize}Serializer".constantize
+  end
+
+  def controlled_resources
+    @controlled_resouces ||= super.where("#{parent_resource}": resource_id)
+  end
+
+  protected
+
+  def set_language_from_header
+    params[:language] = "en"
+  end
+
+  def serializer
+    @serializer ||= "#{resource_name.camelize}Serializer".constantize
+  end
+end
